i am so so sorry but this isnt functional just yet
i am finding this okay to understand, especially as concepts themselves, but am struggling a lot at the same time. i do not know how to explain this but i think its more when it comes to putting it down into practise rather than note-taking on things i am finding it hard to get things done in time or at all in terms of functionality.